pal: svsm: suppor trusted files
This add support of trusted files for svsm. The PAL calculates the hash
values of trusted files when loading and checks if it mathes the value
in the manifest. We can get the sha256 hash value with `sha256sum`
command.

The code (tf_file.c and tf_file.h) is adopted from the gramine-tdx. As
the comment in the original file says, most of the code is the same as
the trusted file handling of linux-sgx' pal and we should unify them.

Example of manifest:
```
[fs]
root.type = "tmpfs"
mounts = [
  { type = "pseudo", path = "/lib", uri = "lib" },
  { path = "/external", uri = "file:/root/module/libwallet/external" },
]

[svsm]
file_check_policy = "strict"
allowed_files = [
  "file:/root/module/libwallet/external/lib",
]
trusted_files = [
  { uri="file:/root/module/libwallet/external/helloworld", sha256="9e7f594364c7d8b3bff5cc4984c7192fce14aa676a9ce145fe3553119b557258"},
  { uri="file:/root/module/libwallet/external/test.txt", sha256="5891b5b522d5df086d0ff0b110fbd9d21bb4fc7163af34d08286a2e846f6be03"},
]
```